Connecting hardware to software
Using arduino to connect hardware(esp32 microcontroller) to collect data and then upload it to the aws cloud. Connected sensors such as moisture sensor and infared sensor to esp32 to test for pests and soil moisture. Using this data along with calculated mathematical formulas we find a plants health score. This information is collected every minute and then uploaded to the aws cloud.